fix: crypto volume parsing and backtest results display
- Change bar.volume from int to float to preserve fractional crypto volumes
- Fix results display in backtest examples to use correct keys from BacktestEngine
- Calculate winning/losing trades from trades list with pnl field

+++ src/gateway/alpaca_data_gateway.py
@@ -145,7 +145,7 @@ class AlpacaDataGateway(DataGateway):
             high=float(alpaca_bar.high),
             low=float(alpaca_bar.low),
             close=float(alpaca_bar.close),
-            volume=int(alpaca_bar.volume),
+            volume=float(alpaca_bar.volume),
             vwap=float(alpaca_bar.vwap) if alpaca_bar.vwap else None,
             trade_count=int(alpaca_bar.trade_count) if alpaca_bar.trade_count else None,
         )
